Write as floats

+++ src/main/scala/pw/byakuren/knuckles/external/APIAnalytics.scala
@@ -23,7 +23,7 @@ class APIAnalytics(botName: String,
         .time(System.currentTimeMillis(), TimeUnit.MILLISECONDS)
         .tag("bot", botName)
         .tag("shard_id", shardId.toString)
-        .addField("value", count)
+        .addField("value", count.toFloat)
         .build()
     )
   }
@@ -34,7 +34,7 @@ class APIAnalytics(botName: String,
         .tag("bot", botName)
         .tag("channel", channelId.toString)
         .tag("guild", guildId.toString)
-        .addField("value", 1)
+        .addField("value", 1.0f)
         .build()
     )
   }
